chore: rename generic/generic_handler to generic/handler
