+++ src/monopoly/generic/__init__.py
@@ -1,4 +1,4 @@
 from .generic import DateMatch, DatePatternAnalyzer
-from .generic_handler import GenericBank, GenericStatementHandler
+from .handler import GenericBank, GenericStatementHandler
 
 __all__ = ["DatePatternAnalyzer", "DateMatch", "GenericStatementHandler", "GenericBank"]
